# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GTA RP Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /servers/{serverId}/channels:
    get:
      operationId: listChannels
      tags: [channels]
      summary: List channels in a server
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Channel"
    post:
      operationId: createChannel
      tags: [channels]
      summary: Create a channel
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelInput"
      responses:
        "201":
          description: Created channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
  /servers/{serverId}/channels/{channelId}:
    get:
      operationId: getChannel
      tags: [channels]
      summary: Get channel
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
    delete:
      operationId: deleteChannel
      tags: [channels]
      summary: Delete channel
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /channels/{channelId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Created message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/{messageId}:
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /characters:
    get:
      operationId: listMyCharacters
      tags: [characters]
      summary: List current user characters
      responses:
        "200":
          description: Characters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Character"
    post:
      operationId: createCharacter
      tags: [characters]
      summary: Create a character
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CharacterInput"
      responses:
        "201":
          description: Created character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
  /characters/{characterId}:
    get:
      operationId: getCharacter
      tags: [characters]
      summary: Get character
      parameters:
        - name: characterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
    patch:
      operationId: updateCharacter
      tags: [characters]
      summary: Update character
      parameters:
        - name: characterId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CharacterUpdate"
      responses:
        "200":
          description: Updated character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
    delete:
      operationId: deleteCharacter
      tags: [characters]
      summary: Delete character
      parameters:
        - name: characterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /dm/{userId}:
    get:
      operationId: getDmMessages
      tags: [directMessages]
      summary: Get DM messages with a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: DM messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendDmMessage
      tags: [directMessages]
      summary: Send a DM to a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, displayName, status, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        status:
          type: string
          enum: [online, idle, dnd, offline]
        rpStatus:
          type: ["string", "null"]
          description: GTA RP in-game status
        activeCharacterId:
          type: ["integer", "null"]
        createdAt:
          type: string
...
    Channel:
      type: object
      required: [id, serverId, name, type, position, createdAt]
      properties:
        id:
          type: integer
        serverId:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [text, voice, announcement, category]
        topic:
          type: ["string", "null"]
        position:
          type: integer
        parentId:
          type: ["integer", "null"]
          description: Parent category channel id
        createdAt:
          type: string
    ChannelInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum: [text, voice, announcement, category]
        topic:
          type: string
        position:
          type: integer
        parentId:
          type: ["integer", "null"]
    Message:
      type: object
      required: [id, content, authorId, author, createdAt]
      properties:
        id:
          type: integer
        content:
          type: string
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/User"
        channelId:
          type: ["integer", "null"]
        dmRecipientId:
          type: ["integer", "null"]
        isSystem:
          type: boolean
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 2000
...
    Character:
      type: object
      required: [id, userId, name, age, occupation, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        age:
          type: integer
        occupation:
          type: string
          description: e.g. LSPD Officer, Mechanic, Drug Dealer, Doctor
        backstory:
          type: ["string", "null"]
        factionId:
          type: ["integer", "null"]
        faction:
          $ref: "#/components/schemas/Faction"
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    CharacterInput:
      type: object
      required: [name, age, occupation]
      properties:
        name:
          type: string
        age:
          type: integer
        occupation:
          type: string
        backstory:
          type: string
        factionId:
          type: ["integer", "null"]
        imageUrl:
          type: string
    CharacterUpdate:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
        occupation:
          type: string
        backstory:
          type: string
        factionId:
          type: ["integer", "null"]
        imageUrl:
          type: string
        isActive:
          type: boolean
    Faction:
      type: object
      required: [id, name, type, color]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [law, medical, criminal, civilian, government, other]
        color:
          type: string
          description: Hex color code for the faction
        description:
          type: ["string", "null"]
        iconUrl:
          type: ["string", "null"]
...